requirements.txt
accelerate==1.11.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.1
aiosignal==1.4.0
anyio==4.11.0
apispec==6.8.4
attrs==25.4.0
blinker==1.9.0
bm25s==0.2.14
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.3
cycler==0.12.1
datasets==4.3.0
dill==0.4.0
evaluate==0.4.6
faiss-cpu==1.12.0
filelock==3.19.1
Flask==3.1.2
flask-smorest==0.46.2
flatbuffers==25.9.23
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2025.9.0
ftfy==6.3.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
humanfriendly==10.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.9
MarkupSafe==2.1.5
marshmallow==4.0.1
matplotlib==3.10.7
mpmath==1.3.0
multidict==6.7.0
multiprocess==0.70.16
networkx==3.5
numpy==2.3.4
onnxruntime==1.23.2
packaging==25.0
pandas==2.3.3
peft==0.17.1
pillow==12.0.0
propcache==0.4.1
protobuf==6.33.0
psutil==7.1.1
pyarrow==21.0.0
pyparsing==3.2.5
pyreadline3==3.5.4
pysbd==0.3.4
PyStemmer==3.0.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
regex==2025.10.23
requests==2.32.5
safetensors==0.6.2
scipy==1.16.2
setuptools==70.2.0
six==1.17.0
sniffio==1.3.1
# Editable install with no version control (spacey-bundle==0.0.3)
# -e d:\proj\releases\ait500g4r # do this with: pip install -e .
sympy==1.14.0
tokenizers==0.22.1
torch==2.9.0+cu129
tqdm==4.67.1
transformers==4.57.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.14
webargs==8.7.0
Werkzeug==3.1.3
xxhash==3.6.0
yarl==1.22.0
